+++ src/service/getGherkinDiagnostics.ts
@@ -1,10 +1,11 @@
 import { Expression } from '@cucumber/cucumber-expressions'
 import { dialects, Errors } from '@cucumber/gherkin'
 import { walkGherkinDocument } from '@cucumber/gherkin-utils'
+import * as messages from '@cucumber/messages'
 import { Diagnostic, DiagnosticSeverity } from 'vscode-languageserver-types'
 
 import { parseGherkinDocument } from '../gherkin/parseGherkinDocument.js'
-import { diagnosticCodeUndefinedStep } from './constants.js'
+import { CONTAINS_PARAMETERS, diagnosticCodeUndefinedStep } from './constants.js'
 
 // https://microsoft.github.io/language-server-protocol/specifications/specification-3-17/#diagnostic
 export function getGherkinDiagnostics(
@@ -48,35 +49,72 @@ export function getGherkinDiagnostics(
   const noStars = (keyword: string) => keyword !== '* '
   const codeKeywords = [...dialect.given, ...dialect.when, ...dialect.then].filter(noStars)
   let snippetKeyword = dialect.given.filter(noStars)[0]
+  let examples: readonly messages.Examples[]
 
   return walkGherkinDocument<Diagnostic[]>(gherkinDocument, diagnostics, {
     scenario(scenario, diagnostics) {
       inScenarioOutline = (scenario.examples || []).length > 0
+      examples = inScenarioOutline ? scenario.examples : []
       return diagnostics
     },
     step(step, diagnostics) {
-      if (inScenarioOutline) {
-        return diagnostics
-      }
       if (codeKeywords.includes(step.keyword)) {
         snippetKeyword = step.keyword
       }
-
-      if (isUndefined(step.text, expressions) && step.location.column !== undefined) {
-        const line = step.location.line - 1
-        const character = step.location.column - 1 + step.keyword.length
-        const diagnostic: Diagnostic = makeUndefinedStepDiagnostic(
-          line,
-          character,
-          step.keyword,
-          step.text,
-          snippetKeyword
-        )
-        return diagnostics.concat(diagnostic)
-      }
-      return diagnostics
+      return inScenarioOutline
+        ? getOutlineStepDiagnostics(step, diagnostics, expressions, snippetKeyword, examples)
+        : getStepDiagnostics(step, diagnostics, expressions, snippetKeyword)
     },
   })
+}
+
+function getOutlineStepDiagnostics(
+  step: messages.Step,
+  diagnostics: Diagnostic[],
+  expressions: readonly Expression[],
+  snippetKeyword: string,
+  examples: readonly messages.Examples[]
+): Diagnostic[] {
+  // Interpolate steps containing parameters
+  if (CONTAINS_PARAMETERS.test(step.text)) {
+    for (const example of examples) {
+      for (const row of example.tableBody) {
+        const stepText = interpolate(
+          step.text,
+          // @ts-ignore Can not be undefined with non-empty table body
+          example.tableHeader.cells,
+          row.cells
+        )
+
+        if (!isUndefined(stepText, expressions)) {
+          return diagnostics
+        }
+      }
+    }
+  }
+
+  return getStepDiagnostics(step, diagnostics, expressions, snippetKeyword)
+}
+
+function getStepDiagnostics(
+  step: messages.Step,
+  diagnostics: Diagnostic[],
+  expressions: readonly Expression[],
+  snippetKeyword: string
+): Diagnostic[] {
+  if (isUndefined(step.text, expressions) && step.location.column !== undefined) {
+    const line = step.location.line - 1
+    const character = step.location.column - 1 + step.keyword.length
+    const diagnostic: Diagnostic = makeUndefinedStepDiagnostic(
+      line,
+      character,
+      step.keyword,
+      step.text,
+      snippetKeyword
+    )
+    return diagnostics.concat(diagnostic)
+  }
+  return diagnostics
 }
 
 export function makeUndefinedStepDiagnostic(
@@ -116,4 +154,22 @@ function isUndefined(stepText: string, expressions: readonly Expression[]): bool
     if (expression.match(stepText)) return false
   }
   return true
+}
+
+function interpolate(
+  name: string,
+  variableCells: readonly messages.TableCell[],
+  valueCells: readonly messages.TableCell[]
+): string {
+  variableCells.forEach((variableCell, n) => {
+    const valueCell = valueCells[n]
+    const valuePattern = '<' + variableCell.value + '>'
+    const escapedPattern = valuePattern.replace(/[-/\\^$*+?.()|[\]{}]/g, '\\$&')
+    const regexp = new RegExp(escapedPattern, 'g')
+    // JS Specific - dollar sign needs to be escaped with another dollar sign
+    // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/replace#Specifying_a_string_as_a_parameter
+    const replacement = valueCell.value.replace(new RegExp('\\$', 'g'), '$$$$')
+    name = name.replace(regexp, replacement)
+  })
+  return name
 }